backend/requirements.txt
flask
flask-cors
mysql-connector-python
flask-jwt-extended
werkzeug
